Add continuous integration with Travis
The Travis build uses gcc-v5 and cmake v3.2.

The build succeeds if test_SatsumaSynteny2 exits successfully. The
output redirected to /dev/null; this is required since Travis does not
like long messages (failing the build if log is greater than 4MB).

NOTE: I had to decrement the required cmake version from 3.3 to 3.2 in
order to be compatible with Travis.